+++ src/cacheddateformat.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999,2004 The Apache Software Foundation.
+ * Copyright 1999,2005 The Apache Software Foundation.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -161,6 +161,6 @@ void CachedDateFormat::setTimeZone(const TimeZonePtr& timeZone) {
   }
 
 
-void CachedDateFormat::numberFormat(LogString& s, long n, Pool& p) const {
+void CachedDateFormat::numberFormat(LogString& s, int n, Pool& p) const {
   formatter->numberFormat(s, n, p);
 }
